# python-flask/swagger_server/models/client.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Client(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, name: str=None, addres: str=None, birth_date: date=None):  # noqa: E501
        """Client - a model defined in Swagger

        :param name: The name of this Client.  # noqa: E501
        :type name: str
        :param addres: The addres of this Client.  # noqa: E501
        :type addres: str
        :param birth_date: The birth_date of this Client.  # noqa: E501
        :type birth_date: date
        """
        self.swagger_types = {
            'name': str,
            'addres': str,
            'birth_date': date
        }

        self.attribute_map = {
            'name': 'name',
            'addres': 'addres',
            'birth_date': 'birthDate'
        }
        self._name = name
        self._addres = addres
        self._birth_date = birth_date

    @classmethod
    def from_dict(cls, dikt) -> 'Client':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Client of this Client.  # noqa: E501
        :rtype: Client
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this Client.


        :return: The name of this Client.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Client.


        :param name: The name of this Client.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def addres(self) -> str:
        """Gets the addres of this Client.


        :return: The addres of this Client.
        :rtype: str
        """
        return self._addres

    @addres.setter
    def addres(self, addres: str):
        """Sets the addres of this Client.


        :param addres: The addres of this Client.
        :type addres: str
        """

        self._addres = addres

    @property
    def birth_date(self) -> date:
        """Gets the birth_date of this Client.


        :return: The birth_date of this Client.
        :rtype: date
        """
        return self._birth_date

    @birth_date.setter
    def birth_date(self, birth_date: date):
        """Sets the birth_date of this Client.


        :param birth_date: The birth_date of this Client.
        :type birth_date: date
        """
        if birth_date is None:
            raise ValueError("Invalid value for `birth_date`, must not be `None`")  # noqa: E501

        self._birth_date = birth_date